+++ symmetric-core/src/main/java/org/jumpmind/symmetric/observability/metrics/HostMetricsService.java
@@ -20,68 +20,22 @@
  */
 package org.jumpmind.symmetric.observability.metrics;
 
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.opentelemetry.api.metrics.DoubleHistogram;
-import io.opentelemetry.api.metrics.LongCounter;
+import io.opentelemetry.api.common.Attributes;
 
 /**
  * This class is intended for collecting a small number of in-memory metrics, which are not engine-specific, but rather describe the host (server) as a whole.
  */
-class HostMetricsService implements IMetricsService {
+class HostMetricsService extends AbstractMetricsService {
 
     private static final Logger log = LoggerFactory.getLogger(HostMetricsService.class);
 
-
-    private final MetricsManager metricsManager;
-    private final Map<String, LongCounter> counters = new ConcurrentHashMap<>();
-    private final Map<String, DoubleHistogram> histograms = new ConcurrentHashMap<>();
-
-    // private HostMetricsService() {
-    // log.debug("Starting Host metrics service");
-    // this.meter = MetricsManager.getGlobalInstance().getOpenTelemetry().getMeter(INSTRUMENTATION_SCOPE);
-    // MetricsManager.getGlobalInstance().register(this);
-    // log.debug("Started Host metrics service");
-    // }
-
-    HostMetricsService(MetricsManager metricsManager) {
-        this.metricsManager = metricsManager;
+    HostMetricsService(MetricsManager metricsManager, boolean isOtelPublishingEnabled) {
+        super(metricsManager, Attributes.empty(), isOtelPublishingEnabled);
         log.debug("Started Host metrics service");
     }
-
-
-    //
-    // @Override
-    // public void createGauge(String name, String description, String unit,
-    // Supplier<Double> valueSupplier) {
-    // meter.gaugeBuilder(name)
-    // .setDescription(description)
-    // .setUnit(unit)
-    // .buildWithCallback(measurement -> measurement.record(valueSupplier.get()));
-    // }
-    //
-    // @Override
-    // public void createCounter(String name, String description, long delta) {
-    // LongCounter counter = counters.computeIfAbsent(name, k ->
-    // meter.counterBuilder(k)
-    // .setDescription(description)
-    // .build());
-    // counter.add(delta);
-    // }
-    //
-    // @Override
-    // public void createHistogram(String name, String description, String unit, double value) {
-    // DoubleHistogram histogram = histograms.computeIfAbsent(name, k ->
-    // meter.histogramBuilder(k)
-    // .setDescription(description)
-    // .setUnit(unit)
-    // .build());
-    // histogram.record(value);
-    // }
 
     @Override
     public void shutdown() {
